# openapi.yaml
openapi: 3.0.0
info:
  title: PB138 Study Planner API
  description: API specification for the PB138 Study Planning Application
  version: 1.0.0
  contact:
    name: PB138 Project
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.pb138.example.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TaskPriority:
      type: string
      enum: [NONE, LOW, MEDIUM, HIGH, URGENT]
      description: Priority level of a task
    Task:
      type: object
      required:
        - id
        - userId
        - title
        - done
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the task
          example: 123
        userId:
          type: integer
          format: int64
          description: ID of the user who owns this task
          example: 1
        title:
          type: string
          description: Task title
          example: "Write essay on history"
          minLength: 1
        done:
          type: boolean
          description: Whether the task is completed
          default: false
          example: false
        subjectId:
          type: integer
          format: int64
          nullable: true
          description: ID of the associated subject, if any
          example: 42
        studyPlanId:
          type: integer
          format: int64
          nullable: true
          description: ID of the associated study plan, if any
          example: 5
        favorite:
          type: boolean
          description: Whether the task is marked as favorite
          default: false
          example: false
        tag:
          type: string
          nullable: true
          description: Custom tag or label for the task
          example: "homework"
        deadline:
          type: string
          format: date-time
          nullable: true
          description: Deadline for completing the task (ISO 8601 format)
          example: "2026-05-15T23:59:59.000Z"
        deletedAt:
          type: string
          format: date-time
          nullable: true
          description: Soft delete timestamp (null if not deleted)
          example: null
    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Task title
          minLength: 1
          example: "Study for exam"
        done:
          type: boolean
          description: Whether the task is completed
          default: false
          example: false
        subjectId:
          type: integer
          format: int64
          nullable: true
          description: ID of the associated subject
          example: 42
        studyPlanId:
          type: integer
          format: int64
          nullable: true
          description: ID of the associated study plan
          example: 5
        favorite:
          type: boolean
          description: Mark as favorite
          default: false
          example: false
        tag:
          type: string
          nullable: true
          description: Custom tag or priority name
          example: "urgent"
        priority:
          $ref: '#/components/schemas/TaskPriority'
          description: Task priority (alternative to tag)
          example: "HIGH"
        deadline:
          type: string
          format: date
          nullable: true
          description: Deadline date (ISO 8601 format)
          example: "2026-05-15"
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          example: "Updated task title"
        done:
          type: boolean
          example: true
        subjectId:
          type: integer
          format: int64
          nullable: true
          example: 42
        studyPlanId:
          type: integer
          format: int64
          nullable: true
          example: 5
        favorite:
          type: boolean
          example: true
        tag:
          type: string
          nullable: true
          example: "updated-tag"
        priority:
          $ref: '#/components/schemas/TaskPriority'
          example: "MEDIUM"
        deadline:
          type: string
          format: date
          nullable: true
          example: "2026-05-20"
    BulkTaskRequest:
      type: object
      required:
        - tasks
      properties:
        tasks:
          type: array
          description: Array of tasks to upsert
          minItems: 1
          items:
            type: object
            required:
              - id
              - title
              - done
            properties:
              id:
                type: integer
                format: int64
                example: 123
              title:
                type: string
                minLength: 1
                example: "Task title"
              done:
                type: boolean
                example: false
              subjectId:
                type: integer
                format: int64
                nullable: true
                example: 42
    PaginatedTasks:
      type: object
      required:
        - data
        - hasMore
        - nextCursor
        - limit
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        hasMore:
          type: boolean
          description: Whether there are more results available
          example: false
        nextCursor:
          type: string
          nullable: true
          description: Cursor for the next page
          example: "456"
        limit:
          type: integer
          description: Number of items returned
          example: 30
...
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
    BulkTasksResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            tasks:
              type: array
              items:
                $ref: '#/components/schemas/Task'
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message describing what went wrong
          example: "Neplatne ID ukolu."
security:
  - bearerAuth: []
paths:
  /api/tasks:
    get:
      summary: List tasks with optional filtering and pagination
      tags:
        - Tasks
      operationId: listTasks
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 30
          description: Number of items per page (max 200)
        - name: cursor
          in: query
          schema:
            type: string
          description: Cursor for pagination (ID of the last item from previous page)
        - name: paginated
          in: query
          schema:
            type: boolean
          description: Enable pagination mode (true/false)
        - name: subjectId
          in: query
          schema:
            type: integer
            format: int64
          description: Filter by subject ID
        - name: studyPlanId
          in: query
          schema:
            type: integer
            format: int64
          description: Filter by study plan ID
        - name: done
          in: query
          schema:
            type: boolean
          description: Filter by completion status (true/false)
        - name: favorite
          in: query
          schema:
            type: boolean
          description: Filter by favorite status (true/false)
        - name: tag
          in: query
          schema:
            type: string
          description: Filter by tag/label
        - name: search
          in: query
          schema:
            type: string
          description: Search in task titles (case-insensitive)
        - name: deadlineFrom
          in: query
          schema:
            type: string
            format: date
          description: Filter tasks with deadline from this date (ISO 8601)
        - name: deadlineTo
          in: query
          schema:
            type: string
            format: date
          description: Filter tasks with deadline until this date (ISO 8601)
        - name: includeDeleted
          in: query
          schema:
            type: boolean
          description: Include soft-deleted tasks (default false)
      responses:
        '200':
          description: List of tasks (with or without pagination)
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Task'
                    description: Regular list when pagination is not enabled
                  - $ref: '#/components/schemas/PaginatedTasks'
                    description: Paginated response when pagination is enabled
              examples:
                unpaginated:
                  value:
                    - id: 1
                      userId: 1
                      title: "Study for exam"
                      done: false
                      subjectId: 42
                      studyPlanId: 5
                      favorite: false
                      tag: "urgent"
                      deadline: "2026-05-15T23:59:59.000Z"
                      deletedAt: null
                paginated:
                  value:
                    data:
                      - id: 1
                        userId: 1
                        title: "Study for exam"
                        done: false
                        subjectId: 42
                        studyPlanId: 5
                        favorite: false
                        tag: "urgent"
                        deadline: "2026-05-15T23:59:59.000Z"
                        deletedAt: null
                    hasMore: true
                    nextCursor: "42"
                    limit: 30
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a new task
      tags:
        - Tasks
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            examples:
              simple:
                summary: Simple task
                value:
                  title: "Read chapter 5"
              detailed:
                summary: Task with all fields
                value:
                  title: "Solve problem set"
                  done: false
                  subjectId: 42
                  studyPlanId: 5
                  favorite: true
                  tag: "homework"
                  deadline: "2026-05-20"
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Bulk update/upsert tasks
      description: |
        Upsert (update or insert) multiple tasks at once.
        - Creates tasks with IDs that don't exist
        - Updates existing tasks
        - Soft-deletes tasks not in the request
      tags:
        - Tasks
      operationId: bulkUpdateTasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkTaskRequest'
            example:
              tasks:
                - id: 1
                  title: "Updated task"
                  done: true
                  subjectId: 42
                - id: 2
                  title: "Another task"
                  done: false
                  subjectId: null
      responses:
        '200':
          description: Tasks upserted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkTasksResponse'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tasks/{id}:
    patch:
      summary: Update a task
      tags:
        - Tasks
      operationId: updateTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Task ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
            example:
              title: "Updated title"
              done: true
              favorite: false
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - invalid ID or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete a task (soft delete)
      description: Soft-deletes a task by setting its deletedAt timestamp
      tags:
        - Tasks
      operationId: deleteTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Task ID
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Bad request - invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...